+++ NeoCom.DataManagement/src/main/java-model/org/dimensinfin/eveonline/neocom/interfaces/IAggregableItem.java
@@ -3,7 +3,7 @@ package org.dimensinfin.eveonline.neocom.interfaces;
 import org.dimensinfin.core.interfaces.ICollaboration;
 
 public interface IAggregableItem extends ICollaboration {
-	long getQuantity();
+	int getQuantity();
 	double getVolume();
 	double getPrice();
 }
